code refactoring - kv cache update
