+++ src/plugins/intel_npu/src/plugin/npuw/kv_cache_sliding_window_manager.cpp
@@ -8,6 +8,7 @@
 #include <limits>
 #include <vector>
 
+#include "infer_request_utils.hpp"
 #include "llm_compiled_model_utils.hpp"
 #include "logging.hpp"
 #include "npuw_transformations/detect_causal_mask.hpp"
@@ -175,6 +176,131 @@ ov::npuw::util::SwaLayout ov::npuw::util::detect_swa_layout(const std::shared_pt
     LOG_DEBUG("[SWA] Layer pattern (S=sliding, F=full attention): " << pattern);
 
     return layout;
+}
+
+void ov::npuw::util::write_swa_kv_slice_circular(ov::SoPtr<ov::ITensor> dst_tensor,
+                                                 ov::SoPtr<ov::ITensor> src_new_kv,
+                                                 uint32_t dst_kv_dim,
+                                                 uint32_t src_kv_dim,
+                                                 uint32_t num_stored_tokens_before,
+                                                 uint32_t num_new_tokens) {
+    const uint32_t capacity = static_cast<uint32_t>(dst_tensor->get_shape()[dst_kv_dim]);
+    const uint32_t old_total = num_stored_tokens_before;
+    const uint32_t new_total = old_total + num_new_tokens;
+    const uint32_t new_valid = std::min(new_total, capacity);
+
+    // Clamp by source length as well: source may already be capacity-limited.
+    const uint32_t src_len = static_cast<uint32_t>(src_new_kv->get_shape()[src_kv_dim]);
+    const uint32_t tokens_to_write = std::min({num_new_tokens, new_valid, src_len});
+
+    if (tokens_to_write == 0) {
+        return;
+    }
+    const uint32_t first_new_abs_pos = num_stored_tokens_before + (num_new_tokens - tokens_to_write);
+    const uint32_t dst_start = first_new_abs_pos % capacity;
+
+    auto src_slice = (src_len > tokens_to_write)
+                         ? ov::npuw::util::make_tensor_slice(src_new_kv, src_kv_dim, src_len - tokens_to_write, src_len)
+                         : src_new_kv;
+
+    if (dst_start + tokens_to_write <= capacity) {
+        auto dst_slice =
+            ov::npuw::util::make_tensor_slice(dst_tensor, dst_kv_dim, dst_start, dst_start + tokens_to_write);
+        ov::npuw::util::copy_tensor_by_dim(src_slice, dst_slice, src_kv_dim, dst_kv_dim);
+    } else {
+        const uint32_t first_leg_len = capacity - dst_start;
+        const uint32_t second_leg_len = tokens_to_write - first_leg_len;
+
+        auto src_first_leg = ov::npuw::util::make_tensor_slice(src_slice, src_kv_dim, 0u, first_leg_len);
+        auto dst_first_leg = ov::npuw::util::make_tensor_slice(dst_tensor, dst_kv_dim, dst_start, capacity);
+        ov::npuw::util::copy_tensor_by_dim(src_first_leg, dst_first_leg, src_kv_dim, dst_kv_dim);
+
+        auto src_second_leg = ov::npuw::util::make_tensor_slice(src_slice, src_kv_dim, first_leg_len, tokens_to_write);
+        auto dst_second_leg = ov::npuw::util::make_tensor_slice(dst_tensor, dst_kv_dim, 0u, second_leg_len);
+        ov::npuw::util::copy_tensor_by_dim(src_second_leg, dst_second_leg, src_kv_dim, dst_kv_dim);
+    }
+}
+
+void ov::npuw::util::write_swa_kv_slice_left_aligned(ov::SoPtr<ov::ITensor> dst_tensor,
+                                                     ov::SoPtr<ov::ITensor> src_new_kv,
+                                                     uint32_t dst_kv_dim,
+                                                     uint32_t src_kv_dim,
+                                                     uint32_t num_stored_tokens_before,
+                                                     uint32_t num_new_tokens) {
+    // Left-aligned SWA policy keeps valid tokens packed at the beginning
+    // After update, logical order to be preserved as:
+    //   [surviving old tail | newest appended tokens], truncated to capacity.
+    const uint32_t capacity = static_cast<uint32_t>(dst_tensor->get_shape()[dst_kv_dim]);
+    const uint32_t old_total = num_stored_tokens_before;
+    const uint32_t new_total = old_total + num_new_tokens;
+    const uint32_t old_valid = std::min(old_total, capacity);
+    const uint32_t new_valid = std::min(new_total, capacity);
+
+    // Clamp against source length too. Some source tensors can hold fewer tokens
+    // than num_new_tokens when they were capacity-limited earlier.
+    const uint32_t src_len = static_cast<uint32_t>(src_new_kv->get_shape()[src_kv_dim]);
+    const uint32_t tokens_to_write = std::min({num_new_tokens, new_valid, src_len});
+
+    // keep: number of old tokens that remain visible after appending the new chunk.
+    // If keep < old_valid, the window is saturated and we must shift surviving old
+    // tokens to the front before appending new ones.
+    const uint32_t keep = new_valid - tokens_to_write;
+    const bool needs_shift = (keep > 0 && keep < old_valid);
+
+    if (needs_shift && dst_kv_dim == 3u) {
+        // Transposed-V (dim=3) partial-slice shifts degrade into many small remote
+        // transfers. Use a full-buffer round-trip to keep transfer count low:
+        //   1) copy full dst -> CPU tmp,
+        //   2) rearrange entirely on CPU,
+        //   3) copy full CPU tmp -> dst.
+        // This avoids many fine-grained remote transactions in the shift phase.
+        LOG_DEBUG("[SWA] Bulk-shifting KV buffer (dim=3): keeping last " << keep << " of " << old_valid
+                                                                         << " old token(s), capacity=" << capacity);
+        auto whole_tmp =
+            ov::npuw::util::allocMem(dst_tensor->get_element_type(), dst_tensor->get_shape(), "CPU", nullptr);
+        dst_tensor->copy_to(whole_tmp._ptr);  // single bulk contiguous transfer
+
+        auto old_tail_cpu = ov::npuw::util::make_tensor_slice(whole_tmp, dst_kv_dim, old_valid - keep, old_valid);
+        auto shift_tmp =
+            ov::npuw::util::allocMem(dst_tensor->get_element_type(), old_tail_cpu->get_shape(), "CPU", nullptr);
+        old_tail_cpu->copy_to(shift_tmp._ptr);  // isolate surviving tail before front overwrite
+        auto dst_front_cpu = ov::npuw::util::make_tensor_slice(whole_tmp, dst_kv_dim, 0u, keep);
+        ov::npuw::util::copy_tensor_by_dim(shift_tmp, dst_front_cpu, dst_kv_dim, dst_kv_dim);
+
+        if (tokens_to_write > 0) {
+            auto src_slice =
+                (src_len > tokens_to_write)
+                    ? ov::npuw::util::make_tensor_slice(src_new_kv, src_kv_dim, src_len - tokens_to_write, src_len)
+                    : src_new_kv;
+            auto dst_back_cpu = ov::npuw::util::make_tensor_slice(whole_tmp, dst_kv_dim, keep, keep + tokens_to_write);
+            ov::npuw::util::copy_tensor_by_dim(src_slice, dst_back_cpu, src_kv_dim, dst_kv_dim);
+        }
+
+        whole_tmp->copy_to(dst_tensor._ptr);  // single bulk contiguous transfer back
+        return;
+    }
+
+    if (needs_shift) {
+        // Window saturated: move the surviving old tail to the front.
+        // Use a temporary buffer to avoid overlapping in-place copy.
+        LOG_DEBUG("[SWA] Shifting KV buffer: keeping last "
+                  << keep << " of " << old_valid << " old token(s), dim=" << dst_kv_dim << ", capacity=" << capacity);
+        auto old_tail = ov::npuw::util::make_tensor_slice(dst_tensor, dst_kv_dim, old_valid - keep, old_valid);
+        auto tmp = ov::npuw::util::allocMem(dst_tensor->get_element_type(), old_tail->get_shape(), "CPU", nullptr);
+        old_tail->copy_to(tmp._ptr);
+        auto dst_front = ov::npuw::util::make_tensor_slice(dst_tensor, dst_kv_dim, 0u, keep);
+        ov::npuw::util::copy_tensor_by_dim(tmp, dst_front, dst_kv_dim, dst_kv_dim);
+    }
+
+    if (tokens_to_write == 0) {
+        return;
+    }
+    // Append newest source tail right after the preserved prefix.
+    auto src_slice = (src_len > tokens_to_write)
+                         ? ov::npuw::util::make_tensor_slice(src_new_kv, src_kv_dim, src_len - tokens_to_write, src_len)
+                         : src_new_kv;
+    auto dst_back = ov::npuw::util::make_tensor_slice(dst_tensor, dst_kv_dim, keep, keep + tokens_to_write);
+    ov::npuw::util::copy_tensor_by_dim(src_slice, dst_back, src_kv_dim, dst_kv_dim);
 }
 
 void ov::npuw::util::fill_causal_sliding_mask(ov::SoPtr<ov::ITensor> mask_tensor,
