+++ src/plugins/intel_npu/src/plugin/npuw/kv_cache_sliding_window_manager.hpp
@@ -77,6 +77,24 @@ void fill_attention_masks(const std::shared_ptr<ov::IAsyncInferRequest>& request
                           uint32_t window_size,
                           const int64_t* token_type_ids_real = nullptr);
 
+// Writes SWA KV deltas into a left-aligned past buffer.
+// If the window is saturated, shifts the surviving tail to the front before append.
+void write_swa_kv_slice_left_aligned(ov::SoPtr<ov::ITensor> dst_tensor,
+                                     ov::SoPtr<ov::ITensor> src_new_kv,
+                                     uint32_t dst_kv_dim,
+                                     uint32_t src_kv_dim,
+                                     uint32_t num_stored_tokens_before,
+                                     uint32_t num_new_tokens);
+
+// Writes SWA KV deltas into a circular past buffer.
+// Token at absolute position p is written to physical slot (p % capacity).
+void write_swa_kv_slice_circular(ov::SoPtr<ov::ITensor> dst_tensor,
+                                 ov::SoPtr<ov::ITensor> src_new_kv,
+                                 uint32_t dst_kv_dim,
+                                 uint32_t src_kv_dim,
+                                 uint32_t num_stored_tokens_before,
+                                 uint32_t num_new_tokens);
+
 }  // namespace util
 }  // namespace npuw
 }  // namespace ov
