+++ src/plugins/intel_npu/tests/unit/npuw/write_kv_slice_sliding_test.cpp
@@ -8,6 +8,7 @@
 #include <vector>
 
 #include "infer_request_utils.hpp"
+#include "kv_cache_sliding_window_manager.hpp"
 #include "openvino/openvino.hpp"
 #include "openvino/runtime/make_tensor.hpp"
